+++ microsoft-graph/release_notes.md
@@ -1,3 +1,11 @@
 #### What's Improved
-- Added support for a certificate-based authentication method. 
-  - Certificate Based Authentication is a new option in the configuration parameter Get Access Token.
+- Security alert-related actions now support API versions V1 and V2. 
+  - Added a new parameter `API Version` with support options V1 and V2 to the following actions
+     - Get All Security Alerts
+     - Get Security Alert
+     - Update Security Alert
+  - Added new parameters `Classification`, `Determination`, `Status`, `Created Date`, `Last Update Date`, `Severity`, and `Service Source` to the `Get All Security Alerts` action based on the selected V2 API Version options.
+  - Added new parameters `Classification`, `Determination`, `Status`,  to the `Update Security Alert` action based on the selected V2 API Version options.
+
+- Added the following new actions and playbooks
+  - Add Comment on Security Alert